% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{get_learner_probs}
\alias{get_learner_probs}
\title{Get probabilities to draw a learner}
\usage{
get_learner_probs(learners, sampling = "parset_dims")
}
\arguments{
\item{learners}{[character(n)] vector of learner id's}

\item{sampling}{[character(1)] "uniform" or "parset_dims".
"uniform" draws each learner equally likely, while parset dims
draws learners proportional to their parameter set dimensions.}
}
\value{
[numeric(n)] sampling probabilities for each learner
}
\description{
Get probabilities to draw a learner
}
