% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasks.R
\name{get_task_metadata}
\alias{get_task_metadata}
\title{Get task metadata}
\usage{
get_task_metadata(oml.task = NULL)
}
\arguments{
\item{oml.task}{[OMLTask|NULL] a OpenML Task.}
}
\value{
[list] named list of metafeatures
}
\description{
Get task metadata
}
