% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{list_learners}
\alias{list_learners}
\title{List available learners for a given task}
\usage{
list_learners(task)
}
\arguments{
\item{task}{[Task] input task}
}
\value{
[character(n)] vector of learner.ids
}
\description{
List available learners for a given task
}
