% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load_or_create_registry}
\alias{load_or_create_registry}
\title{Either loads of creates a registry}
\usage{
load_or_create_registry(regname = "bot_registry", overwrite = FALSE,
  source.pkgs = c("mlr", "data.table"), source.files)
}
\description{
Either loads of creates a registry
}
