% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{make_check_parset}
\alias{make_check_parset}
\title{Create and check a Paramset}
\usage{
make_check_parset(learner, add_fixed_pars, parset, fixed_pars = NULL)
}
\arguments{
\item{learner}{[Learner] object}

\item{parset}{[ParamSet], that matches the learner}
}
\value{
[ParamSet]
}
\description{
Create a learner-parameter set.
  Check whether param set and learner amtch
}
