% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bot.R
\docType{data}
\name{OMLBot}
\alias{OMLBot}
\title{R6-Class OMLBot}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OMLBot
}
\description{
Public Members:
task.id [character(1)]
task [Task]
}
\details{
initialize [function(task.id)]: instantiate a new bot with a given oml task.id
run [function(timeout)] : draw random learner/config and run the bot

Private Members:
learner [learner]
parset [ParamSet]
pars [list]

get_learner_config [function()]
run_internal [function()]
}
\keyword{datasets}
